added reconnection handling

+++ minecraft/startup.lua
@@ -1,13 +1,14 @@
 -- defines software version
 VERSION = 0.30
 
--- defines data types used to identify how data should be interpreted
+-- defines data types used to identify how incoming data should be interpreted
 TYPE_EXEC = "[e]"
 TYPE_CLONE = "[c]"
 TYPE_NAME = "[n]"
 
--- connect to the server
-ws, err = http.websocket("ws://rx-78-2:3000")
+-- defines data types to send data expressions between languages
+TRUE = "{T}"
+FALSE = "{F}"
 
 -- create trust message
 TRUST_MESSAGE = "shake-my-hand-bro"
@@ -21,6 +22,7 @@ function disconnect()
     ws.close()
 end
 
+-- gets the index of an item
 function getItemIndex(itemName)
 	for slot = 1, 16, 1 do
 		local item = turtle.getItemDetail(slot)
@@ -40,43 +42,51 @@ end
 
 -- spawn and boot up another turtle
 function clone()
-    -- places the disk drive
-    turtle.select(getItemIndex("computercraft:disk_drive"))
-    turtle.dig()
-    local step1 = turtle.place()
+    if turtle.getFuelLevel() > 1 then
+        -- places the disk drive
+        turtle.select(getItemIndex("computercraft:disk_drive"))
+        turtle.dig()
+        local step1 = turtle.place()
 
-    -- inserts the disk
-    turtle.select(getItemIndex("computercraft:disk"))
-    local step2 = turtle.drop()
+        -- inserts the disk
+        turtle.select(getItemIndex("computercraft:disk"))
+        local step2 = turtle.drop()
 
-    os.sleep(1)
+        os.sleep(1)
 
-    -- updates the boot file
-    shell.run("rm", "disk/startup.lua")
-    local step3 = shell.run("wget", "https://raw.githubusercontent.com/OGBlackDiamond/turtle-overhaul/main/minecraft/startup.lua", "disk/startup.lua")
+        -- updates the boot file
+        shell.run("rm", "disk/startup.lua")
+        local step3 = shell.run("wget", "https://raw.githubusercontent.com/OGBlackDiamond/turtle-overhaul/main/minecraft/startup.lua", "disk/startup.lua")
 
-    -- updates turtle
-    shell.run("cp", "disk/startup.lua", "startup.lua")
+        -- updates turtle
+        shell.run("cp", "disk/startup.lua", "startup.lua")
 
-    -- moves the turtle up
-    turtle.digUp()
-    local step4 = turtle.up()
+        -- moves the turtle up
+        turtle.digUp()
+        local step4 = turtle.up()
 
-    -- places the new turtle
-    turtle.select(getItemIndex("computercraft:turtle_normal"))
-    turtle.dig()
-    local step5 = turtle.place()
+        -- places the new turtle
+        turtle.select(getItemIndex("computercraft:turtle_normal"))
+        turtle.dig()
+        local step5 = turtle.place()
 
-	turtle.select(1)
-	turtle.drop(math.floor(turtle.getItemCount() / 2))
+        turtle.select(1)
+        turtle.drop(math.floor(turtle.getItemCount() / 2))
 
-    -- turns it on
-    peripheral.call("front", "turnOn")
+        -- turns it on
+        peripheral.call("front", "turnOn")
 
-    -- waits for clone turtle to identify its parent
-    os.sleep(5)
+        -- waits for clone turtle to identify its parent
+        os.sleep(5)
 
-    return step1 and step2 and step3 and step4 and step5
+        if step1 and step2 and step3 and step4 and step5 then
+            return TRUE
+        else
+            return FALSE
+        end
+    end
+    print("not enough fuel to clone")
+    return FALSE
 end
 
 
@@ -86,58 +96,71 @@ end
     This is in a method because it will be called after the parent data has been aquired from its maker
 --]]
 function websocket_start(turtleID, parentID)
-    -- checks if the connection failed
-    if err then
-        print("Failed to connect to the websocket. Power me off and try again.")
-    else
+    -- connect to the server
+    repeat
+        print("attempting websocket connection")
+        ws, err = http.websocket("ws://rx-78-2:3000")
+    until ws ~= false
 
-        -- sends the secondary handshake
-        ws.send(TRUST_MESSAGE)
-        -- print the status of the handshake
-        shake_status = ws.receive()
-        pcall(loadstring(shake_status))
 
-        -- sends the ingame IDs of the two turtles
-        ws.send(turtleID)
-        ws.send(parentID)
+    -- sends the secondary handshake
+    ws.send(TRUST_MESSAGE)
+    -- print the status of the handshake
+    shake_status = ws.receive()
+    pcall(loadstring(shake_status))
 
-        -- MAIN CODE
-        while true do
-            local data = ws.receive(5)
-            local response = nil
-            if data ~= nil then
-                -- parses the data for its type and content
-                local data_type = string.sub(data, 1, 3)
-                local data_content = string.sub(data, 4)
+    -- sends the ingame IDs of the two turtles
+    ws.send(turtleID)
+    ws.send(parentID)
 
-                -- performs the appropriate action based on data type
+    -- MAIN CODE
+    while true do
+        local data = ws.receive(5)
+        local response = nil
+        if data ~= nil then
+            -- parses the data for its type and content
+            local data_type = string.sub(data, 1, 3)
+            local data_content = string.sub(data, 4)
 
-                -- executes the command in data_content
-                if data_type == TYPE_EXEC then
-                    command = loadstring(data_content)
-                    response = command()
+            -- performs the appropriate action based on data type
 
-                -- performs a clone
-                elseif data == TYPE_CLONE then
-                    response = clone()
+            -- executes the command in data_content
+            if data_type == TYPE_EXEC then
+                command = loadstring(data_content)
+                response = command()
 
-                -- sets the name to data_content
-                elseif data_type == TYPE_NAME then
-                    set_name(data_content)
+            -- performs a clone
+            elseif data == TYPE_CLONE then
+                response = clone()
 
-                -- the message recieved did not have a recognizable type
-                else
-                    print("data did not use propper type formatting")
-                end
+            -- sets the name to data_content
+            elseif data_type == TYPE_NAME then
+                set_name(data_content)
 
+            -- the message recieved did not have a recognizable type
+            else
+                print("Unrecognized data formatting")
             end
-            ws.send(response)
+
         end
+        ws.send(response)
     end
 end
 
+-- gets the turtles' id and label
 turtleID = os.getComputerID()
-parentID = peripheral.call("back", "getID")
+turtle_name = os.getComputerLabel()
+
+--[[
+    if the turtle name has not been set, it has never connected to the websocket.
+    it should check behind it for the turtle that spawned it and get it's ID.
+--]]
+if turtle_name == nil then
+    parentID = peripheral.call("back", "getID")
+else 
+    -- if the name is set, we know it has connected, and it will be reconnected
+    parentID = -1
+end
 
 -- copies data to turtle
 shell.run("cp", "/disk/startup.lua", "/startup.lua")
